% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_detailed_report.R
\name{create_detailed_report}
\alias{create_detailed_report}
\title{Function to create a detailed report of the data.}
\usage{
create_detailed_report(default = df)
}
\value{
imports the data from the file and runs the shiny app.
}
\description{
This function uses DataExporer package and create a detailed report of all the data present int the data frame
}
\note{
To use this function by itself, you need to first run import_data() function.Then run the function by writing create_detailed_report()
}
\examples{
create_detailed_report()
}
