% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_func.R
\name{main_func}
\alias{main_func}
\title{Main Function to call sub functions}
\usage{
main_func()
}
\value{
imports the data from the file and runs the shiny app.
}
\description{
This function is called to invoke all the other functions.
}
\examples{
main_func()
}
