% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{Shiny app function.}
\usage{
run_shiny_app()
}
\value{
Shiny app with the titanic data analysis and plots.
}
\description{
This function runs the shiny app to display the analysis.
}
\examples{
run_shiny_app()
}
