% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_mean_class.R
\name{survival_mean_class}
\alias{survival_mean_class}
\title{Relationship between ticket class and survival.}
\usage{
survival_mean_class()
}
\arguments{
\item{df}{The Titanic dataframe which gets imported after running the import_data functions.}

\item{sub_class:}{Subset dataframe containing the Class for passengers and their respective survival rate.}

\item{sboc}{: Variable which stores the the analysis of the mean ticket class}

\item{plot_sboc:}{Variable to store the graph behind the analysis}

\item{sboc_output}{: a variable which stores both the analysis and plot results in a list.}
}
\value{
the analysis behind the relationship between the ticket class and also the graph based on the relationship.
}
\description{
This function analyses the relationship between survival rate and ticket class of the passengers.
}
\examples{
survival_mean_class()
}
