% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_sex_rel.R
\name{survival_sex_rel}
\alias{survival_sex_rel}
\title{Survival rate based on Sex.}
\usage{
survival_sex_rel(default = df)
}
\arguments{
\item{df:}{The titanic train data set which we imported and converted to df dataframe.}

\item{sub_sex}{: subset which contains just the sex and survival rate of passenger}

\item{fem_total}{:total females in the ship}

\item{male_total}{:total males in the ship}

\item{sur_sex}{:subset with only survived passengers}

\item{female_survived}{: total female survived}

\item{male_survived}{: total male survived}

\item{mf_total:}{total analysis}

\item{plot_sex}{: graphical representation of the relationship}
}
\value{
sex_rel_ans : returns the analysis and also the graphical representation.
}
\description{
This function analyses the relationship between Survival rate and sex of the passengers.
}
\note{
To use this function by itself, you need to first run import_data() function to import the  Titanic df from the train.csv file.
}
\examples{
survival_sex_rel()
}
