% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivor_age_mean.R
\name{survivor_age_mean}
\alias{survivor_age_mean}
\title{Survivor Age Mean}
\usage{
survivor_age_mean(default = df)
}
\arguments{
\item{df:}{The titanic train data set which we imported and converted to df dataframe.}

\item{sboa}{Variable to store the output of the analysis.}

\item{plot_age}{: subset to store a dataframe containting age and survival rate.}

\item{graph_age}{: variable containing the graph representing the relationship between age and survival rate}
}
\value{
The mean age of the survivors and the most proobable age of survival and the graphical representation between age and survival.
}
\description{
This function analyses and returns the mean age among the people who survived.Its also gives the age group, which is most likely to survive.
}
\note{
To use this function by itself, you need to first run import_data() function.Then run the functions by writing survivor_age_mean(df)
}
\examples{
survivor_age_mean()
}
