% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic_data}
\alias{titanic_data}
\title{The titanic data set containing 891 passenger data who were present during the titanic disaster.}
\format{An object of class \code{data.frame} with 891 rows and 12 columns.}
\usage{
titanic_data
}
\description{
the data set contains :PassengerID,Survived,Pclass,Name,Sex,SibSp,Parch,Ticket,Fare ,Cabin,Embarked.
}
\keyword{datasets}
