% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Imports Data from csv file and stores it in a dataframe.}
\usage{
import_data()
}
\value{
This function returns the  package data after its created.
}
\description{
This function reads the data from a csv-file and returns the  package data.
}
\details{
This functions downloads the train.csv file from the package to the user's local and creates a dataframe.
}
\examples{
import_data()
}
