% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_mean_class.R
\name{survival_mean_class}
\alias{survival_mean_class}
\title{Relationship between ticket class and survival.}
\usage{
survival_mean_class(default = df)
}
\arguments{
\item{df}{The Titanic dataframe which gets imported after running the import_data functions. Raw data as csv from Kaggle, converted to a R dataframe with the function read.csv.}

\item{sub_class:}{Subset dataframe containing the Class for passengers and their respective survival rate.}

\item{plot_sboc:}{Variable to store the graph behind the analysis}

\item{sboc_output}{: a variable which stores both the analysis and plot results in a list.}
}
\value{
the analysis behind the relationship between the ticket class and also the graph based on the relationship.
}
\description{
This function analyses the relationship between survival rate and ticket class of the passengers.
}
\note{
To use this function by itself, you need to input the Titanic df. The DF can be stored as a variable by using the function import_data().
}
\examples{
survival_mean_class()
}
